Updated read_bytes tests
